+++ ImageClasses/ImagePGM.cpp
@@ -9,7 +9,7 @@
 
 #include "../Filter.h"
 
-ImagePGM::ImagePGM(const unsigned int width, const unsigned int height, const unsigned int maxValue,
+ImagePGM::ImagePGM(const unsigned int width, const unsigned int height, const uint16_t maxValue,
                    std::vector<uint16_t>&& pixelData) :
     Image(ImageType::PGM, width, height, maxValue), pixelData(std::move(pixelData)){
